separated arrows example, changed ball recovery to pressure in the heatmap example
